Move login classes to their own package
