+++ app/src/main/AndroidManifest.xml
@@ -35,10 +35,10 @@
 
         </activity>
         <activity
-            android:name=".LoginActivity"
+            android:name=".components.login.LoginActivity"
             android:windowSoftInputMode="adjustResize">
         </activity>
-        <activity android:name=".LoginWebViewActivity" />
+        <activity android:name=".components.login.LoginWebViewActivity" />
         <activity
             android:name=".MainActivity"
             android:configChanges="orientation|screenSize|keyboardHidden|screenLayout|smallestScreenSize">
