+++ app/src/main/java/com/keylesspalace/tusky/SplashActivity.kt
@@ -18,6 +18,7 @@ package com.keylesspalace.tusky
 import android.content.Intent
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
+import com.keylesspalace.tusky.components.login.LoginActivity
 import com.keylesspalace.tusky.components.notifications.NotificationHelper
 import com.keylesspalace.tusky.db.AccountManager
 import com.keylesspalace.tusky.di.Injectable
